requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
pydantic[dotenv]
pydantic-settings
alembic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
asyncpg
psycopg2-binary
greenlet